android UI fixes after UI changes #3
