+++ android/.browserstack
@@ -1,3 +1,3 @@
 [credentials]
-username = mike_M1rIkt
-key = x2YybDe3qVzH1M6tUpM4
+username = mikepasta_PKlm6D
+key = 6dpEtxQmpxfyWkB789F5
