+++ android/caps.json
@@ -1,11 +1,15 @@
 { "capabilities": {
-  "appium:deviceName": "21f050e03a027ece",
-  "appium:platformName": "Android",
-  "appium:platformVersion": "10",
-  "appium:appPackage": "com.socialsuperstore",
-  "appium:appActivity": "com.socialsuperstore.ui.activity.LauncherActivity",
+  "device" : "Samsung Galaxy S21",
+  "os_version" : "11.0",
+  "project" : "First Python project", 
+  "build" : "browserstack-build-12",
+  "name" : "first_test111",
+  "app_url":"bs://8151df81002753baa7ed8e5ac2dfb9baa428e4b0",
+  "appPackage": "com.socialsuperstore",
+  "appActivity": "com.socialsuperstore.ui.activity.LauncherActivity",    
+  "implicit":12000,
   "autoGrantPermissions": true,
-  "appium:unicodeKeyboard": true,
-  "appium:noReset:": false,
-  "appium:resetKeyboard": true }
+  "unicodeKeyboard": true,
+  "noReset:": false,
+  "resetKeyboard": true }
 }
