+++ android/locators/post_locators.py
@@ -29,11 +29,11 @@ SAVE_BTN_BACKGROUND_ITEMS = "com.socialsuperstore:id/submitLabel"
 
 BACKGROUND_STYLE_AND_SIZE_TAB = "Style & size"
 BACKGROUND_TEXT_COLOUR_TAB = "Text colour"
-ALL_TEXT_STYLES = "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.view.ViewGroup/android.widget.FrameLayout/android.view.ViewGroup/android.widget.FrameLayout[2]/android.view.ViewGroup/android.widget.ScrollView/android.view.ViewGroup/android.view.ViewGroup[2]/androidx.recyclerview.widget.RecyclerView/androidx.appcompat.widget.LinearLayoutCompat"
+ALL_TEXT_STYLES = "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.view.ViewGroup/android.widget.FrameLayout/android.view.ViewGroup/android.widget.FrameLayout[2]/android.view.ViewGroup/android.widget.ScrollView/android.view.ViewGroup/android.view.ViewGroup[2]/androidx.recyclerview.widget.RecyclerView/android.widget.FrameLayout"
 TEXT_SIZE_BAR = "com.socialsuperstore.feature_post_editor:id/textSizeSeekBar"
 BACKGROUND_CROP_BTN = "com.socialsuperstore.feature_post_editor:id/cropBtn"
 BACKGROUND_DELETE_BTN = "com.socialsuperstore.feature_post_editor:id/deleteBtn"
-ALL_TEXT_COLOURS = "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.view.ViewGroup/android.widget.FrameLayout/android.view.ViewGroup/android.widget.FrameLayout[2]/android.view.ViewGroup/android.widget.ScrollView/android.view.ViewGroup/android.view.ViewGroup[2]/androidx.recyclerview.widget.RecyclerView/android.view.ViewGroup"
+ALL_TEXT_COLOURS = "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.view.ViewGroup/android.widget.FrameLayout/android.view.ViewGroup/android.widget.FrameLayout[2]/android.view.ViewGroup/android.widget.ScrollView/android.view.ViewGroup/android.view.ViewGroup[2]/androidx.recyclerview.widget.RecyclerView/android.widget.FrameLayout"
 LIKES_IN_POST = "com.socialsuperstore:id/interactionsLikes"
 COMMENTS_IN_POST = "com.socialsuperstore:id/interactionsComments"
 COMMENTS_INPUT_TEXT_FIELD = "com.socialsuperstore.feature_comments:id/commentEditText"
@@ -45,3 +45,7 @@ MODAL_TITLE_TEXT = "com.socialsuperstore:id/titleTopText"
 CLOSE_MODAL_BTN = "com.socialsuperstore:id/closeButton"
 LET_ME_ADD_PRODUCT_FIRST_BTN = "com.socialsuperstore:id/negativeButton"
 CONTINUE_WITHOUT_PRODUCT_BTN = "com.socialsuperstore:id/positiveButton"
+
+# new post
+ASK_QUESTION_PLUS_MENU = "com.socialsuperstore:id/fabAAQ"
+REC_PRODUCT_PLUS_MENU = "com.socialsuperstore:id/fabRecommendProduct"
