+++ android/pytest.ini
@@ -3,4 +3,4 @@ markers =
     regression: mark a test as regression.
     sanity: mark a test as sanity.
 
-addopts = --driver Appium --variables caps.json -v -s --html=report.html --self-contained-html --disable-pytest-warnings
+addopts = --driver BrowserStack --capability browserName Appium -v -s --html=report.html --self-contained-html --disable-pytest-warnings
